--- a/test/utils_test.exs
+++ b/test/utils_test.exs
@@ -1,0 +1,4 @@
+defmodule UtilsTest do
+  use ExUnit.Case
+  doctest Mighty.Utils
+end