Dramatic performance improvements to `Mighty.Preprocessing.CountVectorizer`. Thanks to @polvalente for pair-programming on this PR.

## Benchmarks

Benchmarks were performed with the following configuration:

```
Operating System: macOS
CPU Information: Apple M1 Max
Number of Available Cores: 10
Available memory: 32 GB
Elixir 1.14.4
Erlang 26.0
JIT enabled: true
```

### Old (Slow) Implementation

```
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s

Benchmarking large_corpus ...
WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
I0000 00:00:1710813088.627743 72846239 tfrt_cpu_pjrt_client.cc:349] TfrtCpuClient created.
Benchmarking unit_tests ...
Calculating statistics...
Formatting results...

Name                   ips        average  deviation         median         99th %
unit_tests            1.90    0.00879 min     ±0.49%    0.00879 min    0.00888 min
large_corpus        0.0146       1.15 min     ±0.00%       1.15 min       1.15 min
```

### New (Faster) Implementation

```
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s

Benchmarking large_corpus ...
WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
I0000 00:00:1710813865.308871 72869116 tfrt_cpu_pjrt_client.cc:349] TfrtCpuClient created.
Benchmarking unit_tests ...
Calculating statistics...
Formatting results...

Name                   ips        average  deviation         median         99th %
unit_tests            1.90         0.53 s     ±0.58%         0.53 s         0.53 s
large_corpus         0.100         9.98 s     ±0.00%         9.98 s         9.98 s
```