@josevalim 2024-03-19
I may have asked this before, but maybe a `tokenizers` has a functionality we could use here?