@acalejos 2024-03-19
> I may have asked this before, but maybe a `tokenizers` has a functionality we could use here?

@josevalim In this particular module for the count / tfidf vectorizers, it is actually unopinionated on the tokenizer. In the benchmark I did, I did in fact use the Tokenizers library for the tokenization function, but the default is just string split.

This allows the user to pass a preprocessing function and a tokenization function.